Move feature flag provider to migration utils and rename vfr to cfr function
