+++ interactive_ai/data_migration/migration/scripts/convert_vfr_videos_to_cfr.py
@@ -6,45 +6,16 @@ import os
 import shutil
 import subprocess
 import tempfile
-from enum import Enum
 
 import cv2
 from bson import ObjectId
 
-from migration.utils import IMigrationScript, MongoDBConnection
+from migration.utils import FeatureFlagProvider, IMigrationScript, MongoDBConnection
 from migration.utils.connection import MinioStorageClient
 
 logger = logging.getLogger(__name__)
 
 BUCKET_NAME_VIDEOS = os.environ.get("BUCKET_NAME_VIDEOS", "videos")
-
-
-class FeatureFlagProvider:
-    @staticmethod
-    def _str2bool(text: str) -> bool:
-        TRUE_SET = {"y", "yes", "t", "true", "on", "1"}
-        FALSE_SET = {"n", "no", "f", "false", "off", "0"}
-        buf_input = text.lower()
-        if buf_input in TRUE_SET:
-            return True
-        if buf_input in FALSE_SET:
-            return False
-        raise ValueError(f"Cannot convert {text} to boolean. Expected one of {TRUE_SET} or {FALSE_SET}")
-
-    @classmethod
-    def is_enabled(cls, feature_flag: str | Enum) -> bool:
-        """
-        Returns whether the provided feature flag is enabled
-
-        :param feature_flag: name of the feature flag to check
-        :return: True if enabled, false otherwise
-        """
-        feature_flag_str = feature_flag.name if isinstance(feature_flag, Enum) else feature_flag
-        feature_flag_value = os.environ.get(feature_flag_str)
-        if feature_flag_value is None:
-            logger.warning(f"Attempting to access undefined flag '{feature_flag_str}'; assuming value False.")
-            return False
-        return cls._str2bool(feature_flag_value)
 
 
 class ConvertVFRVideosToCFR(IMigrationScript):
@@ -53,7 +24,7 @@ class ConvertVFRVideosToCFR(IMigrationScript):
     extract and save every frame, then reconstruct the video using the fps from get_fps method.
     The new constant frame rate video overwrites the old variable framerate video stored in the VideoBinaryRepo.
 
-    This migration needs to extract very frame to ensure existing annotations still align with the video.
+    This migration needs to extract every frame to ensure existing annotations still align with the video.
     """
 
     @classmethod
