+++ interactive_ai/data_migration/migration/utils/__init__.py
@@ -2,12 +2,14 @@
 # LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
 from .connection import MongoDBConnection
+from .feature_flags import FeatureFlagProvider
 from .metadata import ChangesetMetadata
 from .migration_script import IMigrationScript
 from .version_manager import VersionManager
 
 __all__ = [
     "ChangesetMetadata",
+    "FeatureFlagProvider",
     "IMigrationScript",
     "MongoDBConnection",
     "VersionManager",
